app/requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
grpcio==1.56.0
grpcio-tools==1.56.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
protobuf==4.23.3
pydantic==1.10.10
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
StrEnum==0.4.15
typing_extensions==4.7.0
uvicorn==0.22.0
websockets==11.0.3
